update:footer,qrcode,..
