+++ Client/package.json
@@ -30,6 +30,7 @@
     "@tanstack/react-query": "^5.60.2",
     "@types/react-paginate": "^7.1.4",
     "antd": "^5.21.3",
+    "awesome-qr": "^2.1.5-rc.0",
     "axios": "^1.7.7",
     "bootstrap": "^5.3.3",
     "chart.js": "^4.4.4",
