+++ Client/src/component/PersonalPage/MovieTicket.css
@@ -1,0 +1,244 @@
+.ticket-tong{
+    display: grid; /* Thay đổi từ flex sang grid */
+    grid-template-columns: repeat(2, 1fr); /* Chia thành 3 cột */
+    justify-content: center; /* Căn giữa các phần tử nếu cần */
+    margin: 0 auto; /* Giữ các vé cân đối trong container */
+}
+.ticket-container {
+    border: 1px solid #dcdcdc;
+    border-radius: 20px; /* Bo góc mềm mại */
+    padding: 30px;
+    width: 340px;
+    background: linear-gradient(145deg, #1e5799, #4bc3b0); /* Gradient từ xanh dương đậm và xanh ngọc */
+    box-shadow: 0 12px 24px rgba(0, 0, 0, 0.1); /* Bóng đổ nhẹ để tạo chiều sâu */
+    font-family: 'Arial', sans-serif;
+    margin: 60px;
+    transition: all 0.3s ease-in-out;
+}
+.no-tickets-form {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 200px; /* Tùy chỉnh chiều cao */
+    background-color: #f9f9f9;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+  }
+  
+  .no-tickets-message {
+    font-size: 16px;
+    color: #555;
+    text-align: center;
+  }
+  
+/* Header vé */
+.ticket-header {
+    display: flex;
+    gap: 8px;
+    margin-bottom: 12px;
+}
+
+.ticket-format,
+.ticket-sub {
+    background-color: #ff8c00; /* Màu cam sáng */
+    color: #fff;
+    padding: 8px 16px;
+    font-size: 14px;
+    border-radius: 50px;
+    font-weight: bold;
+    text-transform: uppercase;
+}
+
+/* Mã vé */
+.ticket-code {
+    font-size: 18px;
+    font-weight: bold;
+    color: #f39c12; /* Màu vàng sáng nổi bật */
+    margin-bottom: 8px;
+    text-align: center;
+}
+
+/* Tên rạp chiếu phim */
+.ticket-cinema {
+    font-size: 20px;
+    font-weight: bold;
+    color: #fff; /* Màu trắng để nổi bật trên nền tối */
+    margin-bottom: 12px;
+    text-transform: uppercase;
+    text-align: center;
+}
+
+/* Ngày và giờ */
+.ticket-date {
+    font-size: 14px;
+    color: #ecf0f1; /* Màu sáng nhẹ */
+    margin-bottom: 16px;
+    text-align: center;
+}
+
+/* Thông tin vé: Thời gian, phòng, ghế */
+.ticket-info {
+    font-size: 14px;
+    display: flex;
+    gap: 16px;
+    font-weight: 500;
+    justify-content: space-between;
+    color: #fff;
+}
+
+.ticket-info span {
+    background-color: #2c3e50; /* Màu xanh xám đậm để tạo sự tương phản */
+    padding: 6px 14px;
+    border-radius: 8px;
+    color: #fff;
+    font-size: 14px;
+}
+
+/* Phần chi tiết vé */
+.ticket-details {
+    margin-top: 20px;
+}
+
+.ticket-item {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 10px;
+    font-size: 15px;
+}
+
+.item-name {
+    font-weight: 500;
+    color: #ecf0f1; /* Màu sáng để dễ đọc */
+}
+
+.item-quantity,
+.item-price {
+    font-weight: bold;
+    color: #f39c12; /* Màu vàng nổi bật */
+}
+
+/* Tổng tiền */
+.total {
+    font-weight: 700;
+    font-size: 18px;
+    color: #ff6347; /* Màu đỏ cam nổi bật cho tổng tiền */
+    margin-top: 10px;
+    text-align: center;
+}
+.ticket-seat {
+    white-space: nowrap; /* Ngăn xuống dòng */
+    overflow: hidden; /* Ẩn phần nội dung tràn */
+    text-overflow: ellipsis; /* Hiển thị dấu "..." nếu nội dung quá dài */
+    max-width: 100%; /* Giới hạn chiều rộng theo vé */
+  }
+  
+  
+/* QR Code */
+/* QR Code Container */
+.ticket-qr {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 20px;
+}
+
+.qr-wrapper {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 15px;
+    background: #fff; /* Nền trắng */
+    border: 3px solid #f39c12; /* Viền vàng nổi bật */
+    border-radius: 16px;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15); /* Bóng đổ nhẹ */
+}
+
+.qr-image {
+    width: 180px; /* Kích thước QR */
+    height: 180px;
+    object-fit: contain;
+    border-radius: 8px; /* Bo góc QR */
+}
+
+.qr-note {
+    margin-top: 10px;
+    font-size: 14px;
+    color: #555; /* Màu chữ nhẹ */
+    text-align: center;
+    font-style: italic;
+}
+
+.qr-placeholder {
+    font-size: 14px;
+    color: #999; /* Màu chữ mờ */
+    text-align: center;
+    font-style: italic;
+}
+
+
+
+
+/* Định dạng tổng tiền */
+.grand-total {
+    font-weight: bold;
+    font-size: 20px;
+    color: #ff6347; /* Màu đỏ cam để tạo điểm nhấn cho tổng tiền */
+    margin-top: 15px;
+    text-align: center;
+}
+
+
+/* Mobile (dưới 768px) */
+/* Điện thoại (dưới 768px) */
+@media (max-width: 768px) {
+    .ticket-tong {
+        grid-template-columns: 1fr; /* Chỉ hiển thị 1 cột */
+       margin-right: 100px;
+    }
+
+    .ticket-container {
+        width: 90%; /* Chiếm 90% chiều rộng màn hình */
+        padding: 10px; /* Giảm khoảng cách padding bên trong */
+        font-size: 14px; /* Thu nhỏ chữ */
+    }
+
+    .ticket-info {
+        flex-direction: column; /* Chuyển thông tin thành hàng dọc */
+        gap: 8px; /* Giảm khoảng cách giữa các mục */
+    }
+
+    .ticket-info span {
+        font-size: 12px; /* Chữ nhỏ hơn */
+        text-align: center; /* Căn giữa nội dung */
+    }
+
+    .ticket-qr {
+        margin-top: 10px;
+        align-items: center; /* Căn giữa QR code */
+    }
+
+    .ticket-qr .ant-qr {
+        width: 100px; /* Thu nhỏ QR code */
+        height: 100px;
+    }
+
+    .banner-img {
+        height: auto; /* Giữ nguyên tỷ lệ ảnh */
+        max-height: 200px; /* Giảm chiều cao banner */
+    }
+
+    .qr-wrapper {
+        padding: 10px;
+    }
+
+    .qr-image {
+        width: 140px;
+        height: 140px;
+    }
+
+    .qr-note {
+        font-size: 12px;
+    }
+}
