+++ Client/src/component/Support/Support.tsx
@@ -12,7 +12,7 @@ const SupportCenter: React.FC = () => {
           <h1>Trung tâm hỗ trợ</h1>
           <div className='header-gt'>
             <p>
-              Trung tâm hỗ trợ khách hàng của <strong>Moveek</strong> sẽ hỗ trợ các vấn đề liên quan đến <strong>mua vé trực tuyến và thanh toán</strong> tại website moveek.com...<br />
+              Trung tâm hỗ trợ khách hàng của <strong>FlickHive</strong> sẽ hỗ trợ các vấn đề liên quan đến <strong>mua vé trực tuyến và thanh toán</strong> tại website FlickHive.com...<br />
               Các vấn đề khác chúng tôi sẽ không tiếp nhận, vui lòng xem thông tin trên website hoặc liên hệ trực tiếp với rạp để được giải quyết.
             </p>
           </div>
@@ -21,8 +21,8 @@ const SupportCenter: React.FC = () => {
         <section className="contact-info">
           <div className="contact-method">
             <h2>Zalo</h2>
-            <a href="https://zalo.me/moveek" target="_blank" rel="noopener noreferrer">
-              zalo.me/moveek
+            <a href="https://zalo.me/FlickHive" target="_blank" rel="noopener noreferrer">
+              zalo.me/FlickHive
             </a>
             <p>09:00 – 23:00</p>
           </div>
@@ -35,14 +35,14 @@ const SupportCenter: React.FC = () => {
 
           <div className="contact-method">
             <h2>Email</h2>
-            <a href="mailto:hotro@moveek.vn">hotro@moveek.vn</a>
+            <a href="mailto:hotro@FlickHive.vn">hotro@FlickHive.vn</a>
             <p>Phản hồi nhanh nhất có thể</p>
           </div>
 
           <div className="contact-method">
             <h2>Hướng dẫn sử dụng</h2>
-            <a href="https://kb.moveek.com" target="_blank" rel="noopener noreferrer">
-              kb.moveek.com
+            <a href="https://kb.FlickHive.com" target="_blank" rel="noopener noreferrer">
+              kb.FlickHive.com
             </a>
             <p>Cách mua vé, thanh toán...</p>
           </div>
@@ -53,18 +53,18 @@ const SupportCenter: React.FC = () => {
             <h2>Câu hỏi thường gặp</h2>
 
             <div className="faq-item">
-              <h3>Rạp Moveek nằm ở đâu?</h3>
+              <h3>Rạp FlickHive nằm ở đâu?</h3>
               <p>
-                Moveek không phải là rạp phim. Moveek là website thông tin điện ảnh cung cấp các thông tin như: Lịch chiếu,
+                FlickHive không phải là rạp phim. FlickHive là website thông tin điện ảnh cung cấp các thông tin như: Lịch chiếu,
                 Đặt vé phim, Tin tức điện ảnh, Review / Đánh giá phim, dữ liệu phim ảnh...
               </p>
               <p>
-                Kể từ tháng 11/2018, Moveek triển khai bán vé trực tuyến cho các cụm rạp trong cả nước.
+                Kể từ tháng 11/2018, FlickHive triển khai bán vé trực tuyến cho các cụm rạp trong cả nước.
               </p>
             </div>
 
             <div className="faq-item">
-              <h3>Tôi có thể mua vé các cụm rạp nào tại Moveek?</h3>
+              <h3>Tôi có thể mua vé các cụm rạp nào tại FlickHive?</h3>
               <p>Hiện tại bạn đã có thể mua vé các hệ thống rạp:</p>
               <ul>
                 <li><a href="#">Beta Cinemas</a></li>
@@ -76,21 +76,21 @@ const SupportCenter: React.FC = () => {
                 <li><a href="#">Rio Cinemas</a></li>
                 <li><a href="#">Touch Cinema</a></li>
               </ul>
-              <p>Các hệ thống khác, Moveek sẽ kết nối trong thời gian sắp tới.</p>
+              <p>Các hệ thống khác, FlickHive sẽ kết nối trong thời gian sắp tới.</p>
             </div>
 
             {/* Thêm các phần hỏi đáp mới ở đây */}
             <div className="faq-item">
               <h3>Tôi đặt vé trước có được không?</h3>
               <p>
-                Moveek không hỗ trợ đặt vé, bạn cần thanh toán để nhận vé.
+                FlickHive không hỗ trợ đặt vé, bạn cần thanh toán để nhận vé.
               </p>
             </div>
 
             <div className="faq-item">
               <h3>Thanh toán xong thì sao nữa?</h3>
               <p>
-                Sau khi thanh toán thành công bạn sẽ nhận được mã vé qua email và tin nhắn đã chọn. Bạn mang mã vé đến quầy để đổi vé giấy. Nếu không nhận được email / tin nhắn, hãy liên hệ với Moveek để được hỗ trợ ngay nhé.
+                Sau khi thanh toán thành công bạn sẽ nhận được mã vé qua email và tin nhắn đã chọn. Bạn mang mã vé đến quầy để đổi vé giấy. Nếu không nhận được email / tin nhắn, hãy liên hệ với FlickHive để được hỗ trợ ngay nhé.
               </p>
             </div>
 
@@ -104,7 +104,7 @@ const SupportCenter: React.FC = () => {
             <div className="faq-item">
               <h3>Tôi đã bị trừ tiền nhưng không nhận được mã vé?</h3>
               <p>
-                Nếu bạn mua vé ở các cụm rạp Moveek hỗ trợ mua trực tuyến, hãy liên hệ ngay thông qua số hotline, chúng tôi sẽ xử lý trong tối đa 15 phút.<br />
+                Nếu bạn mua vé ở các cụm rạp FlickHive hỗ trợ mua trực tuyến, hãy liên hệ ngay thông qua số hotline, chúng tôi sẽ xử lý trong tối đa 15 phút.<br />
                 Đối với các cụm rạp khác, hãy sang website của rạp hoặc gọi số hotline của rạp để được hỗ trợ.
               </p>
             </div>
